lint added

+++ contracts/resolver/fantom/implementations/aaveV3Resolver.sol
@@ -4,7 +4,6 @@ pragma solidity ^0.8.0;
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 interface IAaveV3DataProvider {
-
     function getReserveConfigurationData(address asset)
         external
         view
@@ -37,9 +36,13 @@ contract Variable {
 }
 
 contract AaveV3Resolver is Variable {
-    function getAvailability( uint256 _route, address[] memory _tokens, uint256[] memory _amounts) public view returns (bool) {
-        require(_route == 9, 'invalid-route');
-        uint length = _tokens.length;
+    function getAvailability(
+        uint256 _route,
+        address[] memory _tokens,
+        uint256[] memory _amounts
+    ) public view returns (bool) {
+        require(_route == 9, "invalid-route");
+        uint256 length = _tokens.length;
         for (uint256 i = 0; i < length; i++) {
             IERC20 token_ = IERC20(_tokens[i]);
             (, , , , , , , , bool isActive, ) = aaveV3DataProvider
